:recycle: allows multiple post migration scripts
