+++ package.json
@@ -11,7 +11,7 @@
     "build": "tsc -p tsconfig.build.json",
     "clean": "rimraf dist/*",
     "db:migrate": "prisma migrate deploy",
-    "db:grant-roles": "node ./dist/prisma/grant-roles.js",
+    "db:post-migrate": "node ./dist/prisma/index.js",
     "dev": "NODE_ENV=development nodemon --watch ./src ./src/index.ts",
     "format": "prettier --write \"**/*.{js,ts,md,json}\" && eslint . --fix --ext .js,.ts",
     "generateEnvDeclaration": "yarn run-script scripts/generateEnvDeclaration.ts",
